[#495] Add write to serial
